+++ sopa/io/__init__.py
@@ -1,8 +1,13 @@
-from .imaging import macsima, phenocycler, hyperion, ome_tif
 from .explorer import write, align
 from .standardize import write_standardized
-from .transcriptomics import merscope, xenium, cosmx
-from .histopathology import wsi, wsi_autoscale
+from .reader.cosmx import cosmx
+from .reader.merscope import merscope
+from .reader.xenium import xenium
+from .reader.macsima import macsima
+from .reader.phenocycler import phenocycler
+from .reader.hyperion import hyperion
+from .reader.utils import ome_tif
+from .reader.wsi import wsi, wsi_autoscale
 from .report import write_report
 
 from ..utils.data import blobs, uniform