+++ sopa/io/report/generate.py
@@ -54,6 +54,14 @@ def _kdeplot_vmax_quantile(values: np.ndarray, quantile: float = 0.95):
 
 
 class SectionBuilder:
+    SECTION_NAMES = [
+        "general_section",
+        "cell_section",
+        "channel_section",
+        "transcripts_section",
+        "representation_section",
+    ]
+
     def __init__(self, sdata: SpatialData):
         self.sdata = sdata
         self.adata = self.sdata.tables.get(SopaKeys.TABLE)
@@ -64,8 +72,6 @@ class SectionBuilder:
         return self.adata.uns[SopaKeys.UNS_KEY].get(key, default)
 
     def general_section(self):
-        log.info("Writing general section")
-
         return Section(
             "General",
             [
@@ -82,8 +88,6 @@ class SectionBuilder:
         )
 
     def cell_section(self):
-        log.info("Writing cell section")
-
         shapes_key, _ = get_boundaries(self.sdata, return_key=True)
         coord_system = get_intrinsic_cs(self.sdata, shapes_key)
 
@@ -111,8 +115,6 @@ class SectionBuilder:
         )
 
     def channel_section(self):
-        log.info("Writing channel section")
-
         image = get_spatial_image(self.sdata)
 
         subsections = [
@@ -154,8 +156,6 @@ class SectionBuilder:
         if not self._table_has(SopaKeys.UNS_HAS_TRANSCRIPTS):
             return None
 
-        log.info("Writing transcript section")
-
         mean_transcript_count = self.adata.X.mean(0).A1
         low_average = mean_transcript_count < LOW_AVERAGE_COUNT
 
@@ -183,8 +183,6 @@ class SectionBuilder:
         return Section("Transcripts", [SubSection("Quality controls", QC_subsubsections)])
 
     def representation_section(self, max_obs: int = 400_000):
-        log.info("Writing representation section")
-
         if self._table_has(SopaKeys.UNS_HAS_TRANSCRIPTS):
             sc.pp.normalize_total(self.adata)
             sc.pp.log1p(self.adata)
@@ -209,11 +207,14 @@ class SectionBuilder:
         )
 
     def compute_sections(self) -> list[Section]:
-        sections = [
-            self.general_section(),
-            self.cell_section(),
-            self.channel_section(),
-            self.transcripts_section(),
-            self.representation_section(),
-        ]
+        sections = []
+
+        for name in self.SECTION_NAMES:
+            try:
+                log.info(f"Writing {name}")
+                section = getattr(self, name)()
+                sections.append(section)
+            except Exception as e:
+                log.warn(f"Section {name} failed with error {e}")
+
         return [section for section in sections if section is not None]